pip_requirements.txt
Shapely==1.5.6
fastkml==0.9
lxml==3.4.0
netCDF4==1.1.4.1
numpy==1.9.1
pygeoif==0.4.1
python-dateutil==2.4.0
requests==2.5.0
six==1.9.0
wsgiref==0.1.2
